+++ RELEASES.rst
@@ -3591,3 +3591,6 @@ Planned
 * Accept unlabelled function statements outside of top level for strict
   functions (using hoist semantics), previously they were rejected with
   a SyntaxError (GH-2213)
+
+* Accept unescaped U+2028 and U+2029 in string literals so that all
+  JSON.stringify() output parses with eval() (ES2019) (GH-2235)